JdbcConnectionConfiguration
